requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0.post1
sqlalchemy>=2.0.23  # Still current, but check again soon
aiohttp>=3.9.4
stripe>=7.9.0
reportlab>=4.0.5
python-multipart>=0.0.7
python-dotenv>=1.0.1
jinja2>=3.1.3
aiosmtplib>=3.0.2
alembic>=1.12.1
pytest>=7.4.4
pytest-asyncio>=0.21.2
celery>=5.3.5
redis>=5.0.2
psutil>=5.9.6
beautifulsoup4>=4.12.3
requests>=2.31.0
cryptography>=41.0.8
flower>=2.0.1
